port most of statik
